+++ components/autofill/core/browser/personal_data_manager_cleaner.cc
@@ -27,13 +27,9 @@ PersonalDataManagerCleaner::PersonalDataManagerCleaner(
       pref_service_(pref_service),
       alternative_state_name_map_updater_(alternative_state_name_map_updater) {
   // Check if profile cleanup has already been performed this major version.
-  is_autofill_profile_cleanup_pending_ =
+  is_autofill_profile_dedupe_pending_ =
       pref_service_->GetInteger(prefs::kAutofillLastVersionDeduped) <
       CHROME_VERSION_MAJOR;
-  DVLOG(1) << "Autofill profile cleanup "
-           << (is_autofill_profile_cleanup_pending_ ? "needs to be"
-                                                    : "has already been")
-           << " performed for this version";
 }
 
 PersonalDataManagerCleaner::~PersonalDataManagerCleaner() = default;
@@ -45,9 +41,7 @@ void PersonalDataManagerCleaner::CleanupDataAndNotifyPersonalDataObservers() {
   // defer the insertion to when the observers are notified.
   if (!alternative_state_name_map_updater_
            ->is_alternative_state_name_map_populated() &&
-      base::FeatureList::IsEnabled(
-          features::kAutofillUseAlternativeStateNameMap) &&
-      is_autofill_profile_cleanup_pending_) {
+      is_autofill_profile_dedupe_pending_) {
     alternative_state_name_map_updater_->PopulateAlternativeStateNameMap(
         base::BindOnce(&PersonalDataManagerCleaner::
                            CleanupDataAndNotifyPersonalDataObservers,
@@ -55,11 +49,16 @@ void PersonalDataManagerCleaner::CleanupDataAndNotifyPersonalDataObservers() {
     return;
   }
 
-  // If sync is enabled for autofill, defer running cleanups until address
-  // sync and card sync have started; otherwise, do it now.
-  if (!personal_data_manager_->IsSyncEnabledFor(
+  // If the user has chosen to sync addresses, wait for sync to start before
+  // performing cleanups. Otherwise do them now.
+  if (!personal_data_manager_->IsUserSelectableTypeEnabled(
           syncer::UserSelectableType::kAutofill)) {
     ApplyAddressFixesAndCleanups();
+  }
+  // If the user has chosen to sync payments, wait for sync to start before
+  // performing cleanups. Otherwise do them now.
+  if (!personal_data_manager_->IsUserSelectableTypeEnabled(
+          syncer::UserSelectableType::kPayments)) {
     ApplyCardFixesAndCleanups();
   }
 
@@ -69,7 +68,10 @@ void PersonalDataManagerCleaner::CleanupDataAndNotifyPersonalDataObservers() {
   personal_data_manager_->NotifyPersonalDataObserver();
 }
 
-void PersonalDataManagerCleaner::SyncStarted(syncer::ModelType model_type) {
+void PersonalDataManagerCleaner::SyncStarted(syncer::ModelType model_type) {}
+
+void PersonalDataManagerCleaner::ApplyAddressAndCardFixesAndCleanups(
+    syncer::ModelType model_type) {
   // The profile de-duplication is run once every major chrome version. If the
   // profile de-duplication has not run for the |CHROME_VERSION_MAJOR| yet,
   // |AlternativeStateNameMap| needs to be populated first. Otherwise,
@@ -78,71 +80,62 @@ void PersonalDataManagerCleaner::SyncStarted(syncer::ModelType model_type) {
   // alternative state name map should be re-populated. This is currently not
   // the case due to the `is_alternative_state_name_map_populated()` check. This
   // state should be reset when sync is disabled, together with
-  // `autofill_profile_sync_started` and `contact_info_sync_started`.
-  autofill_profile_sync_started |= model_type == syncer::AUTOFILL_PROFILE;
-  contact_info_sync_started |= model_type == syncer::CONTACT_INFO;
-  if (autofill_profile_sync_started &&
-      (contact_info_sync_started ||
-       !base::FeatureList::IsEnabled(syncer::kSyncEnableContactInfoDataType)) &&
+  // `autofill_profile_sync_started_` and `contact_info_sync_started_`.
+  autofill_profile_sync_started_ |= model_type == syncer::AUTOFILL_PROFILE;
+  contact_info_sync_started_ |= model_type == syncer::CONTACT_INFO;
+  if (autofill_profile_sync_started_ && contact_info_sync_started_ &&
       !alternative_state_name_map_updater_
            ->is_alternative_state_name_map_populated() &&
-      base::FeatureList::IsEnabled(
-          features::kAutofillUseAlternativeStateNameMap) &&
-      is_autofill_profile_cleanup_pending_) {
+      is_autofill_profile_dedupe_pending_) {
     alternative_state_name_map_updater_->PopulateAlternativeStateNameMap(
-        base::BindOnce(&PersonalDataManagerCleaner::SyncStarted,
-                       weak_ptr_factory_.GetWeakPtr(), model_type));
+        base::BindOnce(
+            &PersonalDataManagerCleaner::ApplyAddressAndCardFixesAndCleanups,
+            weak_ptr_factory_.GetWeakPtr(), model_type));
     return;
   }
 
   // Run deferred autofill address profile startup code.
-  if (model_type == syncer::AUTOFILL_PROFILE)
+  if (autofill_profile_sync_started_ && contact_info_sync_started_ &&
+      (model_type == syncer::AUTOFILL_PROFILE ||
+       model_type == syncer::CONTACT_INFO)) {
     ApplyAddressFixesAndCleanups();
+  }
 
   // Run deferred credit card startup code.
-  if (model_type == syncer::AUTOFILL_WALLET_DATA)
+  if (model_type == syncer::AUTOFILL_WALLET_DATA) {
     ApplyCardFixesAndCleanups();
+  }
 }
 
 void PersonalDataManagerCleaner::ApplyAddressFixesAndCleanups() {
-  // One-time fix, otherwise NOP.
-  RemoveOrphanAutofillTableRows();
+  if (!is_profile_cleanup_pending_) {
+    return;
+  }
 
   // Once per major version, otherwise NOP.
   ApplyDedupingRoutine();
 
-  // Once per major version, otherwise NOP.
+  // Ran once on browser startup.
   DeleteDisusedAddresses();
 
-  // Ran everytime it is called.
-  ClearProfileNonSettingsOrigins();
-
-  // Once per user profile startup.
+  // Ran once on browser startup.
   RemoveInaccessibleProfileValues();
+
+  is_profile_cleanup_pending_ = false;
 }
 
 void PersonalDataManagerCleaner::ApplyCardFixesAndCleanups() {
-  // Once per major version, otherwise NOP.
+  if (!is_credit_card_cleanup_pending_) {
+    return;
+  }
+
+  // Ran once on browser startup.
   DeleteDisusedCreditCards();
 
-  // Ran everytime it is called.
+  // Ran once on browser startup.
   ClearCreditCardNonSettingsOrigins();
-}
 
-void PersonalDataManagerCleaner::RemoveOrphanAutofillTableRows() {
-  // Don't run if the fix has already been applied.
-  if (pref_service_->GetBoolean(prefs::kAutofillOrphanRowsRemoved))
-    return;
-
-  scoped_refptr<AutofillWebDataService> local_db =
-      personal_data_manager_->GetLocalDatabase();
-  if (!local_db)
-    return;
-
-  local_db->RemoveOrphanAutofillTableRows();
-
-  // Set the pref so that this fix is never run again.
-  pref_service_->SetBoolean(prefs::kAutofillOrphanRowsRemoved, true);
+  is_credit_card_cleanup_pending_ = true;
 }
 
 void PersonalDataManagerCleaner::RemoveInaccessibleProfileValues() {
@@ -168,22 +161,13 @@ void PersonalDataManagerCleaner::RemoveInaccessibleProfileValues() {
 }
 
 bool PersonalDataManagerCleaner::ApplyDedupingRoutine() {
-  if (!base::FeatureList::IsEnabled(
-          features::kAutofillEnableProfileDeduplication)) {
-    return false;
-  }
-
   // Check if de-duplication has already been performed on this major version.
-  if (!is_autofill_profile_cleanup_pending_) {
-    DVLOG(1)
-        << "Autofill profile de-duplication already performed for this version";
+  if (!is_autofill_profile_dedupe_pending_) {
     return false;
   }
 
   const std::vector<AutofillProfile*>& profiles =
       base::FeatureList::IsEnabled(
-          features::kAutofillAccountProfilesUnionView) &&
-              base::FeatureList::IsEnabled(
                   features::kAutofillAccountProfileStorage)
           ? personal_data_manager_->GetProfiles()
           : personal_data_manager_->GetProfilesFromSource(
@@ -192,14 +176,14 @@ bool PersonalDataManagerCleaner::ApplyDedupingRoutine() {
   // No need to de-duplicate if there are less than two profiles.
   if (profiles.size() < 2) {
     DVLOG(1) << "Autofill profile de-duplication not needed.";
+    pref_service_->SetInteger(prefs::kAutofillLastVersionDeduped,
+                              CHROME_VERSION_MAJOR);
     return false;
   }
 
   DVLOG(1) << "Starting autofill profile de-duplication.";
   std::unordered_set<std::string> profiles_to_delete;
   profiles_to_delete.reserve(profiles.size());
-  // Used to update credit card's billing addresses after the dedupe.
-  std::unordered_map<std::string, std::string> guids_merge_map;
 
   // `profiles` contains pointers to the PDM's state. Modifying them directly
   // won't update them in the database and calling `PDM:UpdateProfile()`
@@ -209,7 +193,7 @@ bool PersonalDataManagerCleaner::ApplyDedupingRoutine() {
     new_profiles.push_back(std::make_unique<AutofillProfile>(*profile));
   }
 
-  DedupeProfiles(&new_profiles, &profiles_to_delete, &guids_merge_map);
+  DedupeProfiles(&new_profiles, &profiles_to_delete);
 
   // Apply the profile changes to the database.
   for (const auto& profile : new_profiles) {
@@ -222,9 +206,7 @@ bool PersonalDataManagerCleaner::ApplyDedupingRoutine() {
     }
   }
 
-  UpdateCardsBillingAddressReference(guids_merge_map);
-
-  is_autofill_profile_cleanup_pending_ = false;
+  is_autofill_profile_dedupe_pending_ = false;
   // Set the pref to the current major version.
   pref_service_->SetInteger(prefs::kAutofillLastVersionDeduped,
                             CHROME_VERSION_MAJOR);
@@ -234,18 +216,13 @@ bool PersonalDataManagerCleaner::ApplyDedupingRoutine() {
 
 void PersonalDataManagerCleaner::DedupeProfiles(
     std::vector<std::unique_ptr<AutofillProfile>>* existing_profiles,
-    std::unordered_set<std::string>* profiles_to_delete,
-    std::unordered_map<std::string, std::string>* guids_merge_map) const {
+    std::unordered_set<std::string>* profiles_to_delete) const {
   AutofillMetrics::LogNumberOfProfilesConsideredForDedupe(
       existing_profiles->size());
 
-  // Sort the profiles by ranking score with all the verified profiles at the
-  // end. That way the most relevant profiles will get merged into the less
-  // relevant profiles, which keeps the syntax of the most relevant profiles
-  // data. Verified profiles are put at the end because they do not merge into
-  // other profiles, so the loop can be stopped when we reach those. However
-  // they need to be in the vector because an unverified profile trying to merge
-  // into a similar verified profile will be discarded.
+  // Sort the profiles by ranking score. That way the most relevant profiles
+  // will get merged into the less relevant profiles, which keeps the syntax of
+  // the most relevant profiles data.
   // Since profiles earlier in the list are merged into profiles later in the
   // list, `kLocalOrSyncable` profiles are placed before `kAccount` profiles.
   // This is because local profiles can be merged into account profiles, but not
@@ -260,13 +237,10 @@ void PersonalDataManagerCleaner::DedupeProfiles(
         }
         return a->HasGreaterRankingThan(b.get(), comparison_time);
       });
-  auto first_verified_profile = base::ranges::stable_partition(
-      *existing_profiles, [](const std::unique_ptr<AutofillProfile>& profile) {
-        return !profile->IsVerified();
-      });
 
   AutofillProfileComparator comparator(personal_data_manager_->app_locale());
-  for (auto i = existing_profiles->begin(); i != first_verified_profile; i++) {
+  for (auto i = existing_profiles->begin(); i != existing_profiles->end();
+       i++) {
     AutofillProfile* profile_to_merge = i->get();
 
     // If the profile was set to be deleted, skip it. This can happen because
@@ -305,8 +279,6 @@ void PersonalDataManagerCleaner::DedupeProfiles(
       // and will not accept updates from `profile_to_merge`.
       if (existing_profile.SaveAdditionalInfo(
               *profile_to_merge, personal_data_manager_->app_locale())) {
-        guids_merge_map->emplace(profile_to_merge->guid(),
-                                 existing_profile.guid());
         profiles_to_delete->insert(profile_to_merge->guid());
 
         // Account profiles track from which service they originate. This allows
@@ -326,13 +298,10 @@ void PersonalDataManagerCleaner::DedupeProfiles(
         // Now try to merge the new resulting profile with the rest of the
         // existing profiles.
         profile_to_merge = &existing_profile;
-        // Verified profiles do not get merged. Save some time by not
-        // trying. Note that the `existing_profile` (now `profile_to_merge`)
-        // might be verified.
-        // Similarly, account profiles cannot be merged into other profiles,
-        // since that would delete them.
-        if (profile_to_merge->IsVerified() ||
-            profile_to_merge->source() == AutofillProfile::Source::kAccount) {
+        // Account profiles cannot be merged into other profiles, since that
+        // would delete them. Note that the `existing_profile` (now
+        // `profile_to_merge`) might be verified.
+        if (profile_to_merge->source() == AutofillProfile::Source::kAccount) {
           break;
         }
       }
@@ -340,62 +309,6 @@ void PersonalDataManagerCleaner::DedupeProfiles(
   }
   AutofillMetrics::LogNumberOfProfilesRemovedDuringDedupe(
       profiles_to_delete->size());
-}
-
-void PersonalDataManagerCleaner::UpdateCardsBillingAddressReference(
-    const std::unordered_map<std::string, std::string>& guids_merge_map) {
-  /*  Here is an example of what the graph might look like.
-
-      A -> B
-             \
-               -> E
-             /
-      C -> D
-  */
-
-  std::vector<CreditCard> server_cards_to_be_updated;
-  for (auto* credit_card : personal_data_manager_->GetCreditCards()) {
-    // If the credit card is not associated with a billing address, skip it.
-    if (credit_card->billing_address_id().empty())
-      continue;
-
-    // If the billing address profile associated with the card has been merged,
-    // replace it by the id of the profile in which it was merged. Repeat the
-    // process until the billing address has not been merged into another one.
-    std::unordered_map<std::string, std::string>::size_type nb_guid_changes = 0;
-    bool was_modified = false;
-    auto it = guids_merge_map.find(credit_card->billing_address_id());
-    while (it != guids_merge_map.end()) {
-      was_modified = true;
-      credit_card->set_billing_address_id(it->second);
-      it = guids_merge_map.find(credit_card->billing_address_id());
-
-      // Out of abundance of caution.
-      if (nb_guid_changes > guids_merge_map.size()) {
-        NOTREACHED();
-        // Cancel the changes for that card.
-        was_modified = false;
-        break;
-      }
-    }
-
-    // If the card was modified, apply the changes to the database.
-    if (was_modified) {
-      if (credit_card->record_type() == CreditCard::LOCAL_CARD)
-        personal_data_manager_->GetLocalDatabase()->UpdateCreditCard(
-            *credit_card);
-      else
-        server_cards_to_be_updated.push_back(*credit_card);
-    }
-  }
-
-  // In case, there are server cards that need to be updated,
-  // |PersonalDataManager::Refresh()| is called after they are updated.
-  if (server_cards_to_be_updated.empty())
-    personal_data_manager_->Refresh();
-  else
-    personal_data_manager_->UpdateServerCardsMetadata(
-        server_cards_to_be_updated);
 }
 
 bool PersonalDataManagerCleaner::DeleteDisusedAddresses() {
@@ -409,17 +322,11 @@ bool PersonalDataManagerCleaner::DeleteDisusedAddresses() {
     return true;
   }
 
-  std::unordered_set<std::string> used_billing_address_guids;
-  for (CreditCard* card : personal_data_manager_->GetCreditCards()) {
-    if (!card->IsDeletable()) {
-      used_billing_address_guids.insert(card->billing_address_id());
-    }
-  }
-
+  // Don't call `PDM::RemoveByGUID()` directly, since this can invalidate the
+  // pointers in `profiles`.
   std::vector<std::string> guids_to_delete;
   for (AutofillProfile* profile : profiles) {
-    if (profile->IsDeletable() &&
-        !used_billing_address_guids.count(profile->guid())) {
+    if (profile->IsDeletable()) {
       guids_to_delete.push_back(profile->guid());
     }
   }
@@ -427,28 +334,12 @@ bool PersonalDataManagerCleaner::DeleteDisusedAddresses() {
   size_t num_deleted_addresses = guids_to_delete.size();
 
   for (auto const& guid : guids_to_delete) {
-    personal_data_manager_
-        ->RemoveAutofillProfileByGUIDAndBlankCreditCardReference(guid);
-  }
-
-  if (num_deleted_addresses > 0) {
-    personal_data_manager_->Refresh();
+    personal_data_manager_->RemoveByGUID(guid);
   }
 
   AutofillMetrics::LogNumberOfAddressesDeletedForDisuse(num_deleted_addresses);
 
   return true;
-}
-
-void PersonalDataManagerCleaner::ClearProfileNonSettingsOrigins() {
-  // `kAccount` profiles don't store an origin.
-  for (AutofillProfile* profile : personal_data_manager_->GetProfilesFromSource(
-           AutofillProfile::Source::kLocalOrSyncable)) {
-    if (profile->origin() != kSettingsOrigin && !profile->origin().empty()) {
-      profile->set_origin(std::string());
-      personal_data_manager_->UpdateProfileInDB(*profile, /*enforced=*/true);
-    }
-  }
 }
 
 void PersonalDataManagerCleaner::ClearCreditCardNonSettingsOrigins() {
